requirements.txt
pytube>=15.0.0
yt-dlp>=2024.3.10
tqdm==4.66.1
requests>=2.31.0
cryptography>=41.0.0
streamlit>=1.32.0
streamlit-extras>=0.3.5
pillow>=10.2.0 